Commit 185418af (open) by Claas, 2023-11-25 00:16
ship_traffic_generator.generate_traffic_situations() : Corrected an error I introduced with commit 185418af on 2023-11-25 00:16
lat_lon_0 of the new traffic_situation needs to be set to lat_lon_0 from the encounter settings. It was, though, set to lat_lon_0 from the desired traffic situation (which is usually None). I introduced the error when changing the code to use the new pdantic data types.
Now corrected. Also added a related assertion to the tests to ensure the error does not get introduced again.